Clean up the test files
